Add files via upload
This file includes the jupyter notebook that includes EDA and sentiment analysis of the dataset that is used in this repository. Sentiment analysis on title and tags and it shows how title and tags impact the performance of the video in views, likes, comment region. And also engagement rate impact .